--- NomadNotifierReact/src/routes/Login.jsx
@@ -1,46 +1,0 @@
-import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import '../styles/Login.css'; // Make sure this path is correct
-
-export default function LoginPage() {
-    const [username, setUsername] = useState('');
-    const [password, setPassword] = useState('');
-    const navigate = useNavigate();
-
-    const handleLogin = (e) => {
-        e.preventDefault();
-        console.log('Login:', username, password);
-    };
-
-    const handleCreateAccount = () => {
-        navigate('/create-account');
-    };
-
-    return (
-        <div className="login-container">
-            <h2>Login</h2>
-            <form onSubmit={handleLogin} className="login-form">
-                <div>
-                    <label htmlFor="username">Username:</label>
-                    <input
-                        type="text"
-                        id="username"
-                        value={username}
-                        onChange={(e) => setUsername(e.target.value)}
-                    />
-                </div>
-                <div>
-                    <label htmlFor="password">Password:</label>
-                    <input
-                        type="password"
-                        id="password"
-                        value={password}
-                        onChange={(e) => setPassword(e.target.value)}
-                    />
-                </div>
-                <button type="submit">Enter</button>
-            </form>
-            <button onClick={handleCreateAccount} className="create-account-btn">Create New Account</button>
-        </div>
-    );
-}
