+++ NomadNotifierReact/src/routes/Travel.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { useNavigate,  useLocation } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 
 
 export default function PlanTravel() {
